feat: api 모듈 변경 및 entity -> document
