+++ Backend/api/src/main/java/com/example/api/ApiApplication.java
@@ -1,10 +1,10 @@
-package com.example;
+package com.example.api;
 
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {"com.example"})
 public class ApiApplication {
     public static void main(String[] args) {
         SpringApplication.run(ApiApplication.class, args);
