+++ Backend/persistence/mongodb/src/main/java/com/example/mongodb/repository/CartRepository.java
@@ -1,9 +1,9 @@
 package com.example.mongodb.repository;
 
-import com.example.mongodb.entity.CartEntity;
+import com.example.mongodb.document.CartDocument;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartRepository extends MongoRepository<CartEntity, String> {
+public interface CartRepository extends MongoRepository<CartDocument, String> {
 }
